3% performance gains by changing Type enum from uint8_t to uint64_t

+++ src/brainfuck/types.h
@@ -12,7 +12,7 @@ enum class Error : int {
   UNMATCHED_BRACE,
 };
 
-enum Type : uint8_t {
+enum Type : uint64_t {
   NOOP,
   DONE,
   DATA_ADD,                             // + // -
